requirements.txt
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
github3.py==1.0.0a4
idna==2.6
matplotlib==2.1.2
numpy==1.14.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.22
